pyproject.toml
[project]
name = "DisMu"
version = "0.1.0"
description = "A Discord Music Bot with comprehensive audio playback capabilities"
authors = [
    {name = "Giovanbattista Abbate"}
]
license = {text = "GPL-3.0"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py~=2.5.2",
    "PyNaCl~=1.5.0",
    "python-dotenv~=1.1.1",
    "yt_dlp"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"